+++ game_patch/rf/geometry.h
@@ -476,11 +476,8 @@ namespace rf
     static auto& g_cache_clear = addr_as_ref<void()>(0x004F0B90);
     static auto& g_get_room_render_list = addr_as_ref<void(GRoom ***rooms, int *num_rooms)>(0x004D3330);
 
-    static auto& g_num_geomod_shapes = addr_as_ref<int>(0x00637220);
-    static auto& g_geomod_shapes_strings = addr_as_ref<String[10]>(0x006371F0);
     static auto& g_solid_load_v3d_embedded = addr_as_ref<GSolid*(const char*)>(0x00586E70);
     static auto& g_solid_load_v3d = addr_as_ref<GSolid*(const char*)>(0x00586F5C);
-    static auto& g_geomod_shapes_meshes = addr_as_ref<int[10]>(0x006371F8);
 
     static auto& material_find_impact_sound_set = addr_as_ref<ImpactSoundSet*(const char* name)>(0x004689A0);